refactor(web): recast ACP agent list as list/detail navigation
- Replace inline card with a list/detail two-view pattern using SwapTransition
- List view: each profile is a row with Switch toggle + ChevronRight to navigate into setup
- Detail view: full configuration form for the selected agent, with a back button
- Handle loading (Skeleton), empty (Empty), and vanished-profile states
- Remove old monolithic settings-acp-card.vue
